requirements.txt
python-telegram-bot
python-dotenv
loguru
langchain
langchain-google-genai
langchain-groq
google-generativeai
duckduckgo-search
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
schedule
PyPDF2
python-docx
openpyxl
pillow
aiohttp
